UberMongo/requirements.txt
pymongo>=4.7
python-dotenv>=1.0
